# Refactor: Represent suggestion results as sparse arrays