--- a/annif/backend/ensemble.py
+++ b/annif/backend/ensemble.py
@@ -3,7 +3,6 @@
 
 import annif.eval
 import annif.parallel
-import annif.suggestion
 import annif.util
 from annif.exception import NotSupportedException
 from annif.suggestion import SuggestionBatch