--- a/annif/backend/mllm.py
+++ b/annif/backend/mllm.py
@@ -9,7 +9,6 @@ import annif.eval
 import annif.util
 from annif.exception import NotInitializedException, NotSupportedException
 from annif.lexical.mllm import MLLMModel
-from annif.suggestion import VectorSuggestionResult
 
 from . import backend, hyperopt
 
@@ -144,7 +143,7 @@ class MLLMBackend(hyperopt.AnnifHyperoptBackend):
         vector = np.zeros(len(self.project.subjects), dtype=np.float32)
         for score, subject_id in prediction:
             vector[subject_id] = score
-        return VectorSuggestionResult(vector)
+        return vector
 
     def _suggest(self, text, params):
         candidates = self._generate_candidates(text)