@osma 2023-03-17
Rebased on current `main` and force-pushed, mainly because I need to consider the CLI refactoring done in #675 before making changes to CLI-related code.